fix(frontend): adjust input fields and parsing for usia and BMI
- Change usia input step from 0.1 to 1 and update placeholder example
- Modify BMI input step from 0.01 to 1 and update placeholder and hint text
- Update JavaScript parsing to use integer instead of float for usia and BMI values

+++ frontend/js/app.js
@@ -717,8 +717,8 @@ function selectRadio(el, name) {
 function initPrediksiForm() {
   document.getElementById('prediksiForm').addEventListener('submit', async (e) => {
     e.preventDefault();
-    const usia = parseFloat(document.getElementById('inputUsia').value);
-    const bmi = parseFloat(document.getElementById('inputBmi').value);
+    const usia = parseInt(document.getElementById('inputUsia').value);
+    const bmi = parseInt(document.getElementById('inputBmi').value);
     const sistolik = parseInt(document.getElementById('inputSistolik').value);
     const diastolik = parseInt(document.getElementById('inputDiastolik').value);
     const genderEl = document.querySelector('input[name="gender"]:checked');
